+++ Options/CHANGELOG.lua
@@ -3,6 +3,9 @@ if not TMW then return end
 TMW.CHANGELOG_LASTVER="10.0.0"
 
 TMW.CHANGELOG = [==[
+## v12.0.5
+* Fixed an issue that could cause some edit mode dialogs to be really wide.
+
 ## v12.0.4
 
 ### Midnight Aura Improvements